pyproject.toml
[project]
name = "mqtt-weather"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.1.0",
    "requests>=2.32.0",
    "rich>=13.9.0",
    "python-dotenv>=1.0.0",
    "meshtastic>=2.5.0",
]
